feat(backend): implement cart endpoints and fix CartProductDTO types

+++ server/src/main/java/com/plugin/server/dto/CartProductDTO.java
@@ -1,12 +1,14 @@
 package com.plugin.server.dto;
 
+import java.math.BigDecimal;
+
 public class CartProductDTO {
     private String name;
-    private double price;
+    private BigDecimal price;
     private String mainImg;
     private int amount;
 
-    public CartProductDTO(String name, double price, String mainImg, int amount) {
+    public CartProductDTO(String name, BigDecimal price, String mainImg, int amount) {
         this.name = name;
         this.price = price;
         this.mainImg = mainImg;
@@ -17,7 +19,7 @@ public class CartProductDTO {
         return name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
